Run prettier as part of the CI pipeline
This will fail CI on any code that does not pass the prettier linter

This also switches over to using the prettier linter as an eslint plugin since the experience directly running it is pretty terrible. It won't show the actual codeblock that is causing an error and always just tells you to use the fix option, making it extremely hard to track things down when the linter fails.

Signed-off-by: Robert Clark <rbclark@mitre.org>