Migrate to the Heimdall2 backend (mitre/heimdall-lite#277)
Migrate to the Heimdall2 backend

This commit takes a lot of the logic that has been copy-pasted between different files and moves it all to a global location.
Initial page load is now paused until we figure out what mode we are in (since we have vue-router wait until we know before it proceeds). We check whether or not `/server` responds with 200 or 404, if it responds with 200 then we know we are in server mode and proceed accordingly.

Before this is merged, the backend.ts Vuex store should be renamed to server.ts and the current server.ts should be removed (once all the functionality has been ported). backend.ts is intended to eventually be a full-featured replacement for server.ts.

* Fix logout function, (CLEAR_LOGOUT is a function not a property)

* Require authorization on every route other than login and signup for Heimdall Server

* Remove unreferenced files, remove all console.log statements

* Turn vue-analytics debug mode off

* Modify declaration of Vuex store and reuse the declaration everwhere.

This gets rid of all calls to getModule in the application and instead places the const declaration in their appropriate store files.

Run all SVG assets through https://jakearchibald.github.io/svgomg/. This is suggested by https://vue-svg-loader.js.org/faq.html#how-to-use-this-loader-with-jest which is the plugin we use for loading SVGs, and fixes issues we will encounter in the test suite when trying to load components that use SVGs.

* Update UI for ece testing

* fix success message

* fix message

* add id to login and error message

* Add support for SVG files in the test suite

Without this change, SVG files raise an error on import during mounting in the test suite.

* Update versions of ts-jest, jest, and typescript to all be within the supported versions for one another

Reorder jest moduleFileExtensions into the order of most to least used.

Fix tests to work with updated getModule calls and mock out the required parts of the Vuex store to test the Topbar.

* use camel case

* add id

* Add additional HTML ids and update existing HTML ids for Login.vue.

* Add HTML ids on switch to login button on register page and the username/email field

* Correct label for password in Signup

* Add back in id for topbar

* Do not call the logout function when no token is set

When app logout is performed the app is normally reloaded. This adds a check for token existence, and if no token exists then the app is on the login page.

Signed-off-by: Robert Clark <rbclark@mitre.org>

* Move File Toggling up to the UploadNexus level

Rename application title to Heimdall (in preparation for Heimdall Server & Lite)

Fix SVG display

Fix clear button in Results.vue

Remove broken tests (all of the ones removed were in the wrong spec file and need to be rewritten in their proper respective spec files)

Signed-off-by: Robert Clark <rbclark@mitre.org>

* Fix awkward div formatting on Login and Signup forms

Co-authored-by: sjoshi10 <sjoshi@mitre.org>
Co-authored-by: sjoshi10 <saurabjdc@gmail.com>
Co-authored-by: Kyle Fagan <kfagan@mitre.org>