Convert all Vue components to use vue-property-decorator declarations
Previously the application was declaring props separately for every component. This made it very annoying touse with typescript since it limited to only primitive types. Now that the application is using vue-property-decorator, it is possible to put any type on a model, making declarations much simpler in a lot of cases.