Fix results and profile view to only display their respective files (#369)
A recent commit broke the Results view causing it to display both results and profiles. This reverts that change.

This also deduplicates the search input element so that the same one is displayed whether the application is in mobile view or full size view.

Also fixes a warning from ProfileData that was being displayed on the Developer Console.