Cleanup Status Counts
This clarifies the status counts along the top of Heimdall, removes the total count, and makes the count of passing and failing individual controls more clear in the failed box. It also stops the counting of "FailedOutOf", "NotApplicableTests", "NotReviewedTests", "ErroredOutOf", "ErroredTests", and "TotalTests" since those counts are no longer used anywhere and counting them would be a waste of CPU cycles.

Closes #80