Added the changes required to make trusted publishing work with npmjs.
Followed documentation from here: https://docs.npmjs.com/trusted-publishers

Changes included 1) the changes to permissions, 2) the addition of the
step to manually upgrade npm to an appropriate version, 3) the removal
of the environment variable being set for the npm publish commands when
publishing to npmjs.

Please note that my understanding is that tokens are still required for
publishing to GitHub packages, but that uses the GITHUB_TOKEN not the
NPM_TOKEN so after testing to confirm that releases still work, we
should be clear to delete NPM_TOKEN secret entirely.

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>