Npm trusted publishing (#7318)
* updated action versions

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* Added the changes required to make trusted publishing work with npmjs.

Followed documentation from here: https://docs.npmjs.com/trusted-publishers

Changes included 1) the changes to permissions, 2) the addition of the
step to manually upgrade npm to an appropriate version, 3) the removal
of the environment variable being set for the npm publish commands when
publishing to npmjs.

Please note that my understanding is that tokens are still required for
publishing to GitHub packages, but that uses the GITHUB_TOKEN not the
NPM_TOKEN so after testing to confirm that releases still work, we
should be clear to delete NPM_TOKEN secret entirely.

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* explicitly set permissions for all workflows

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* try using 3.1 instead of 2.7 for building inspecjs

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* use cinc-auditor instead of inspec

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* I don't want this output to be quiet

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* make script use cinc-auditor

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* quiet implied exit code.  i want the output to actually show up but also the exit code so now we are explicitly calling exit code

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

* something changed with one of the underlying libraries I think and now it's outputting floats instead of ints which is fine since impact is a float from [0,1] anyways

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>

---------

Signed-off-by: Amndeep Singh Mann <amann@mitre.org>